pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hydro-param"
version = "0.1.0dev0"
description = "Configuration-driven hydrologic parameterization for any model, any fabric."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Rich McDonald", email = "rich@connectedwatersllc.com" },
]
keywords = [
    "hydrology",
    "parameterization",
    "geospatial",
    "NHM",
    "NextGen",
    "watershed",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Hydrology",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "xarray",
    "geopandas",
    "numpy",
    "pyyaml",
    "pydantic>=2.0",
]

[project.optional-dependencies]
gdp = ["gdptools", "exactextract"]
stac = ["pystac-client", "planetary-computer", "rioxarray"]
regrid = ["xesmf", "rioxarray"]
parallel = ["joblib"]
cloud = ["coiled", "dask-geopandas"]
download = ["pynhd"]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "pre-commit",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/rmcd-mscb/hydro-param"
Documentation = "https://github.com/rmcd-mscb/hydro-param/blob/main/docs/design.md"
Repository = "https://github.com/rmcd-mscb/hydro-param"
Issues = "https://github.com/rmcd-mscb/hydro-param/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/hydro_param"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]

[[tool.mypy.overrides]]
module = [
    "geopandas.*",
    "shapely.*",
    "xarray.*",
    "numpy.*",
    "pandas.*",
    "pynhd.*",
    "gdptools.*",
    "pystac_client.*",
    "planetary_computer.*",
    "rioxarray.*",
    "exactextract.*",
]
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

# ---------------------------------------------------------------------------
# Pixi workspace configuration
# ---------------------------------------------------------------------------
[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "win-64"]

[tool.pixi.pypi-dependencies]
hydro-param = { path = ".", editable = true }

# Conda-forge overrides for packages with C/Fortran dependencies.
# These replace the PyPI versions from [project.dependencies].
[tool.pixi.dependencies]
geopandas = ">=1.0"
xarray = ">=2024.1"
numpy = ">=1.26"

# --- Features ---------------------------------------------------------------
# Each [project.optional-dependencies] group is auto-detected as a feature.
# Below we add conda overrides and Python-version pins as additional features.

[tool.pixi.feature.gdp.dependencies]
exactextract = ">=0.2"

[tool.pixi.feature.stac.dependencies]
pystac-client = ">=0.8"
planetary-computer = "*"
rioxarray = ">=0.15"

[tool.pixi.feature.regrid.dependencies]
xesmf = ">=0.8"
rioxarray = ">=0.15"

[tool.pixi.feature.download.dependencies]
pynhd = ">=0.17"

[tool.pixi.feature.download.tasks]
download-delaware = "python scripts/download_delaware_nhd.py"

[tool.pixi.feature.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
mypy = "*"
pre-commit = "*"

[tool.pixi.feature.dev.tasks]
test = "pytest tests/ --cov=src/hydro_param --cov-report=term-missing"
lint = "ruff check src/ tests/"
format = "ruff format src/ tests/"
format-check = "ruff format --check src/ tests/"
typecheck = "mypy src/"
pre-commit = "pre-commit run --all-files"
check = { depends-on = ["lint", "format-check", "typecheck", "test"] }

[tool.pixi.feature.py311.dependencies]
python = "3.11.*"

[tool.pixi.feature.py312.dependencies]
python = "3.12.*"

# --- Environments ------------------------------------------------------------
# Pin default/dev/full to Python 3.11 for stability. Separate solve-groups
# for different Python versions to avoid conflicts.
[tool.pixi.environments]
default = { features = ["py311"], solve-group = "py311" }
dev = { features = ["dev", "py311"], solve-group = "py311" }
test-py311 = { features = ["dev", "py311"], solve-group = "py311" }
test-py312 = { features = ["dev", "py312"], solve-group = "py312" }
full = { features = ["dev", "gdp", "regrid", "parallel", "stac", "py311"], solve-group = "py311" }
download = { features = ["download", "py311"], solve-group = "py311" }
